don't use MixedCaps in package name but lowercase only
