+++ systemmonitor/sysMetrics.go
@@ -28,7 +28,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package systemMonitor
+package systemmonitor
 
 import (
 	"bufio"
